Make section table similar to readelf
Change presentation of section headers to format that is same as from
readelf -S option. There are still some differences because of
unsupported processor-specific section types.

In this commit the output table is now generated by a custom print
function instead of a generic automatically-generated algorithm to
tailor to the desired output style.