Don't fail if -p received unknown section name
Original implementation would fail if option --string-dump would receive
a name of unknown section. This commit changes this behavior to be the
same as in readelf - same warning message will be printed and execution
will go on. This resolves some of the issues with test scripts.

Note that the behavior would still differ with readelf if section
numbers would be used instead of section names. But this scenario is not
used by the test scripts, and currently there is no goal to be exactly
identical to readelf.