Add a prototype of call frame instructions support
It is quite clunky and need improvements. Perhaps it should be done as
part of adding support frames interpetation, seens more logic will be
added to this class and it is a good oportunity to think how to tie this
together. Current code is mostly just to read and print instructions.