Support frame interpreation
There is an issue with some frames, which override the definition from
the CIE initial instructions - register rule from initial instructions
should be dropped completely in this case, but currently it is
superseded only when the new instruction is made. This causes difference
when comparing with objdump output for some frames, for example those
that redefine where RA is stored.