Use CallFrameTable to evaluate PC in dwarf=frames printing
The printing format should include PC values, so the code that just
prints call frame instructions still should handle instructions that
change location. Original implementation used a special code for that,
separate from CallFrameTable, causing some duplication. This commit
changes that to use an existing CallFrameTable and providing an access
to necessary data, decreasing code duplication.

One disadvantage of this approach is that now all instructions are
interpreted even for the -dwarf=frames format, causing unnecessary code
execution. In theory this can be mitigated by extracting the
location-tracking code into a separate class, used by CallFrameTable and
by dwarf=frames directly, but this is an unnecessary complication for
this simple project.