Support zero terminator CIE records
Adds support for zero-terminator CIE records in the .eh_frame stream.
Update parsing algorithm with streams that don't have a zero terminator
at the end - for example object files, or executables built without a
startup files.

Replace code which tries to read FDE, and if that fails resets cursor
and tries to reread it as CIE, possible a zero CIE, with a more common
sense approach - first read fields that are common to CIE and FDE:
length and cie_ptr, and then call CIE or FDE reader depending on the
cie_ptr value.