Parse CFI/FDE instructions immediately
Original implementation read CFI and FDE instructions as bytes and
parsed them into instructions only when needed - for example to print
or to build the call frame table. However this approach is problematic
for eh_frame with augmentations where instructions can for example
contain addresses relative to the instruction own offset. To properly
parse those addresses it would be needed to store offset of the
instruction byte buffer itself. Instead I think it is better idea to
just parse instructions immediately when parsing the CFI/FDE itself,
when the offset can be easily known.

This change doesn't yet pass any of those additional offsets yet, it
just provides infrastructure to allow passing of offsets usaful.