Support augmentation-specified pointers for DW_CFA_set_loc
In .eh_frame encoding and adjustment of the pointer argument to
DW_CFA_set_loc depends on information in the augmentation. This commit
adds support for this, although as through the rest of the code only the
pcrel adjustment is supported, as for now I have only test cases that
use this adjustment.