Don't pass CIE record mapping to FDE read function
This is not really needed - the FdeRecord.read function can already
accept the CieRecord object as an argument. The only issue with new
approach is that CIE pointer is still passed as a separate argument, and
in theory it is possible that CIE pointer will point to a different CIE
than the one passed as an argument.