Rename .eh_frame-specific functions
Rename functions to clearly identify that they are specific to .eh_frame
or really for parsing with augmentations. Versions of functions for
DWARF .debug_frame would be named differently, and that would be
different implementations. In theory it would be nice to have just
single implementation (augmentations can occur in DWARF frames as well,
after all), however it seems to me that at least with current
implementation properly supporting augmentations requires passing around
a lot of additional information which makes for bad code design. At this
moment I would prefer to do something else rather than try to solve that
problem and so there would be two implementation for commonly found
.eh_frame and for .debug_frame.