Add a simple test for .debug_abbrev
Initially the idea was to test different attributes and forms, however
because of the nature of .debug_abbrev it doesn't really test anything
except for whether specific attribute or form has been added to
respective enum type. To test the handling of those attributes it is
required to have a complete .debug_info section. In the next commit I
will start adding tests for .debug_info that don't depend on other
sections.