Remove workaround for a binutils bug
There was a binutils bug in generation of call frame tables that
required a workaround in dwarf.py.  Now that it has been fixed the
workaround can be removed, though I haven't dig down to which commit
fixed the problem.  At the time of implementing workaround I wasn't sure
if this was a bug in binutils, or I just don't understand something.