Support abbreviation numbers that are not in orderly sequence
Original implementation assumed that abbreviation numbers are always in
an ordered sequence and without any gaps, which is typical for generated
numbers, however is not a requirement in the specification. This commit
adds support for the case when abbreviation numbers are not an
array-like sequence. The test case debug_info.S already contains such
case.