Fix handling of abbreaviation declaration
The implementation was not correct: it didn't stopped at the end of
compilation unit abbreviations and continued to parse abbreviations of
the next CU. Also the handling of levels for abbreviations is
meaningless: while `has/hasn't` child is stored in the abreviation, the
value applies only to debug information entries, not to abbreviations.
This fixes errors found with by the tests/ folder.